Don't include item and itemScope in collection component
Fixes an error when trying to print the full collection into console or
trying to use the `for ... of` construct.

The problem was that we were generating an invalid component `item`
inside the collection component that was throwing an `Element not found`
error when inspected. By removing this invalid element, the collection
component can be inspected safely.