Update to new async/await unit test syntax
This updates the private unit test code to use the new async/await syntax for testing. The tests are equivalent to what they used to be, but this syntax will allow running the tests against the new rfc268 test API as well.

Other than purely syntactic changes, this also involved a couple of semantic changes:

* Replace `visit()` stubbing with actual calls to `visit()` because there's isn't a great way to stub `visit()` in the new test framework
* Replace `andThen()`/`wait()` adapter calls with an `await()` call. The `await` call is invoked like this: `await this.adapter.await(someHelper())`. In integration tests, it's a no-op and in acceptance tests it ignores the return value of the helper and just awaits the `wait()` helper (making it equivalent to the previous `andThen()/wait()` invocations). But in the new test framework, it will be implement to `await` the return value of the helper, which is how async waiting in the new test framework works.