Fix compatibility exports
When using `require()`, we're not guaranteed that the module we're `requiring` is resolved yet, so it could just be an empty object with no exports. This has actually happened to a user, although it's tightly tied to their module dependency graph, and I haven't been able to easily repro it, nor do I think writing a unit test for it is feasible :( But the fix is to "lazy call" the `@ember/test-helpers` methods so we can be assured that they are resolved by the time we call them, even if they aren't when we export our compatibility functions.