[feat] Allow native getter syntax
This PR allows users to use native `get` syntax directly on page
objects, as opposed to the `getter` macro. This syntax is much less
verbose and easier to read overall.

The implementation first ensures that all property descriptors are
correctly copied to the final definition, then updates objects in the
Ceibo builder to use the equivalent of the getter macro. This could also
be added directly to Ceibo, but would likely still require some changes
to ember-cli-page-object to ensure that descriptors are copied via
`getOwnPropertyDescriptor` and `defineProperty` and not via simple
assignment (which will activate the getter when attempting to assign
it).