Decouple chainability from execution contexts
This commit frees up execution contexts from knowing about chainability
by extracting the logic to the standalone `run` function which
generalizes chainability / asyncronous behavior.

However, in order to keep backward compatibility for existing legacy `moduleFor`
tests, a new `andThen` hook was introduced for the execution contexts.
If the exectution context has one, `run` follows the old actions style, when each
action is waiting for the completeness of the previous action despite we use
`await` or not.

This `andThen` hook should be removed in the next major release, in order to
get trully unified, execution-context agnostic asyncronous behavior.