Decouple chainability from execution contexts
With introducing RFC269 extecution context we've got a solid way to
manage chainability. However, that required to us add chainability aware
code to the legacy exection contexts.

This commit frees up exectution contexts from knowing about chainability
by extracting the logic to the standalone `run` function which
generalizes chainability/asyncrony behavior.

However, in order to keep backward compatibility for existing legacy `moduleFor`
tests, a new `andThen` hook was introduced for execution contexts. If the
exectution context has one, `run` follows the old actions style, when each
action is waiting for the completeness of the previos action  despite we use
`await` or not.

This `andThen` hook should be removed in the next major release, in order to
get full control over asyncrony despite the test helpers used in the exectution
context.