Install ec-ts
It uses ec-typescript@1, cause more recent versions of ec-typescipt
require ember-cli-babel@7 or higher. And that requires node@6
or higher, which is technically a breaking change, cause we
support node@4.5 in v1

However, we use blueprint setup from ec-ts@3 here to have the best
we can for now.

The other changes:
  - Setup build to fail on TS error
  - remove dtslint