stop using rfc268 adapter if not set
While it feels natural to auto-select an offcial RFC268 test mode,
if no adapter was specified; it adds some extra internal complexity,
because of eager import of the `@ember/test-helpers`.
In theory, it's possible to not have `@ember/test-helpers` installed,
which is demonstrated in "with-ember-test-helpers" ember-try scenario.

To solve this we could use `window.require(`, to lazily import needed helpers,
only in case if `@ember/test-helpers` installed. But this aproach may leads
to some issues in Node.js envs, since `window.require` is not a thibng there.
Also I'm not sure if `window.require(` is supposed to work fine with Emberoider.

So, we go an explicit way for now, and require to import,
and set a desired adapter on the test suite side manually.