fix: avoid renaming function on import
For some reason, the Ember Babel plugin isn’t working when we re-name the `deprecate` function while importing; it results in the code being converted into

`Ember.emberDeprecate`

Rather than

`Ember.deprecate`

And that function is, understandably, not defined.

I thought the issue might be the dated version of `ember-cli-babel` but updating that doesn’t seem to actually solve the problem.