Install ember@4 for embroider test scenarios
This is a workaround.

npm@8 hoists a peer dep from transitive deps of the `/addon` folder.
It makes `@embroider/macros` to misbehave and to work against ember version which is not equal to the ember version specified  in test-app .
In its turn, it makes `@ember/test-helpers` to fails.

So by tweaking the ember-source in ember-try scenario,
we allow ember/test-helpers to survive in this deps collision