fix: restore getter error details (#631)
* fix: restore getter error details

(#596) broke QUnit error messages so they now miss contextual error
info like page object path and selector.

This happened cause #596 relied on the `.toString()` method of the Error
to build the final error message. However, that's not how QUnit is
displaying the error message. It just uses the `Error.message`.

So let's build the final message in the `Error` constructor.